c_idx cast to LongTensor in random sparse projection

+++ anomalib/models/components/dimensionality_reduction/random_projection.py
@@ -82,7 +82,7 @@ class SparseRandomProjection:
                 )
                 data = torch.distributions.Binomial(total_count=1, probs=0.5).sample(sample_shape=c_idx.size()) * 2 - 1
                 # assign data to only those columns
-                components[i, c_idx] = data.double()
+                components[i, c_idx.long()] = data.double()
 
             components *= np.sqrt(1 / density) / np.sqrt(self.n_components)
